+++ Fontend/.eslintrc.json
@@ -9,6 +9,6 @@
         "sourceType": "module"
     },
     "rules": {
-    
+
     }
 }
